requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
aiosqlite>=0.19.0

# PDF Processing Libraries
PyPDF2>=3.0.1
reportlab>=4.0.4

# Utilities
pathlib2>=2.3.7
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0